+++ Components/Core/Conditions/Conditions.lua
@@ -636,19 +636,25 @@ function CNDT:GetUnit(setting)
 end
 
 
-function CNDT:GetTableSubstitution(tbl)
-	TMW:ValidateType("CNDT:GetTableSubstitution(tbl)", "tbl", tbl, "table")
+local tableVarMap = {}
+local tableIdx = 1
+-- actually pointless for keys to be GCable because the values are never evicted from Env, but ¯\_(ツ)_/¯
+setmetatable(tableVarMap, { __mode="k" }) 
 
-	-- We used to check the format of the address explicitly,
-	-- but ticket 1076 demonstrates that sometimes it can be in the format
-	-- "table: 0000000312EBEB0" (the format I get), or "table: 0x1ba1bbb00" (from the ticket)
-	-- so instead we just check that the metatable exists.
-	local address = tostring(tbl)
-	if TMW.approachTable(tbl, getmetatable, "__tostring") then
-		error("can't substitute tables with __tostring metamethods: " .. address)
+--- Add the specified table as a variable to CNDT.Env so that var can be used in funcstrs.
+-- @param the table to variablize.
+-- @param suffix An optional suffix useful for debugging and for the variable substitution process
+function CNDT:GetTableSubstitution(tbl, suffix)
+	if tableVarMap[tbl] then
+		return tableVarMap[tbl]
 	end
 
-	local var = address:gsub(":", "_"):gsub(" ", "")
+	TMW:ValidateType(1, "CNDT:GetTableSubstitution(tbl[, suffix])", tbl, "table")
+	TMW:ValidateType(2, "CNDT:GetTableSubstitution(tbl[, suffix])", suffix, "string;nil")
+
+	local var = "table_" .. tableIdx .. (suffix and ("_" .. suffix) or "")
+	tableVarMap[tbl] = var
+	tableIdx = tableIdx + 1
 	CNDT.Env[var] = tbl
 
 	return var
@@ -713,7 +719,7 @@ CNDT.Substitutions = {
 		TMW:ValidateType("c.BitFlags", conditionData.identifier, conditionSettings.BitFlags, "table;number")
 
 		if type(conditionSettings.BitFlags) == "table" then
-			return CNDT:GetTableSubstitution(conditionSettings.BitFlags)
+			return CNDT:GetTableSubstitution(conditionSettings.BitFlags, "BitFlags")
 		elseif type(conditionSettings.BitFlags) == "number" then
 			return conditionSettings.BitFlags
 		end
@@ -737,22 +743,16 @@ CNDT.Substitutions = {
 
 {	src = "c.Level",
 	rep = function(conditionData, conditionSettings, name, name2)
-		TMW:ValidateType("c.Level", conditionData.identifier, conditionSettings.Level, "number")
-
 		return conditionData.percent and conditionSettings.Level/100 or conditionSettings.Level
 	end,
 },{
 	src = "c.Checked",
 	rep = function(conditionData, conditionSettings, name, name2)
-		TMW:ValidateType("c.Checked", conditionData.identifier, conditionSettings.Checked, "boolean")
-
 		return tostring(conditionSettings.Checked)
 	end,
 },{
 	src = "c.Checked2",
 	rep = function(conditionData, conditionSettings, name, name2)
-		TMW:ValidateType("c.Checked2", conditionData.identifier, conditionSettings.Checked2, "boolean")
-
 		return tostring(conditionSettings.Checked2)
 	end,
 },{
@@ -769,18 +769,10 @@ CNDT.Substitutions = {
 
 {
 	src = "c.NameFirst2",
-	rep = function(conditionData, conditionSettings, name, name2)
-		return strWrap(TMW:GetSpells(name2).First)
-
-	end,
+	rep = "error('Condition sub c.NameFirst2 is obsolete. Use c.Spells2.FirstS')",
 },{
 	src = "c.NameString2",
-	rep = function(conditionData, conditionSettings, name, name2)
-		return strWrap(TMW:GetSpells(name2).FirstString)
-	end,
-},{
-	src = "c.ItemID2",
-	rep = "error('Condition sub c.ItemID is obsolete')",
+	rep = "error('Condition sub c.NameString2 is obsolete. Use c.Spells2.FirstString')",
 },{
 	src = "c.Name2Raw",
 	rep = function(conditionData, conditionSettings, name, name2)
@@ -795,22 +787,16 @@ CNDT.Substitutions = {
 
 {
 	src = "c.NameFirst",
+	-- deprecated (for c.Spells.First), but not removed because it is definitely used externally (TMW_ST)
 	rep = function(conditionData, conditionSettings, name, name2)
 		return strWrap(TMW:GetSpells(name).First)
 	end,
 },{
-	src = "c.NameStrings",
-	rep = function(conditionData, conditionSettings, name, name2)
-		return strWrap(";" .. table.concat(TMW:GetSpells(name).StringArray, ";") .. ";")
-	end,
-},{
 	src = "c.NameString",
+	-- deprecated (for c.Spells.FirstString), but not removed because it may be used externally
 	rep = function(conditionData, conditionSettings, name, name2)
 		return strWrap(TMW:GetSpells(name).FirstString)
 	end,
-},{
-	src = "c.ItemID",
-	rep = "error('Condition sub c.ItemID is obsolete')",
 },{
 	src = "c.NameRaw",
 	rep = function(conditionData, conditionSettings, name, name2)
@@ -821,54 +807,48 @@ CNDT.Substitutions = {
 	rep = function(conditionData, conditionSettings, name, name2)
 		return strWrap(name)
 	end,
-},{
-	src = "c.Spells",
-	rep = function(conditionData, conditionSettings, name, name2)
-		return CNDT:GetTableSubstitution(TMW:GetSpells(name))
-	end,
 },
 
 {
-	src = "c.True",
-	rep = 	function(conditionData, conditionSettings, name, name2)
-		return tostring(conditionSettings.Level == 0)
-	end,
-},{
-	src = "c.False",
+	src = "c%.Spells(2?)(%.?[A-Za-z]*)",
 	rep = function(conditionData, conditionSettings, name, name2)
-		return tostring(conditionSettings.Level == 1)
-	end,
-},{
-	src = "c.1nil",
-	rep = function(conditionData, conditionSettings, name, name2)
-		return conditionSettings.Level == 0 and 1 or "nil"
-	end,
-},{
-	src = "c.nil1",
-	rep = function(conditionData, conditionSettings, name, name2)
-		-- reverse 1nil
-		return conditionSettings.Level == 1 and 1 or "nil"
+		return function(two, suffix)
+			local input = two == "2" and name2 or name
+			local spells = TMW:GetSpells(input)
+			if suffix:sub(1,1) == "." then
+				suffix = suffix:trim(".")
+				local subItem = spells[suffix]
+				if subItem then
+					if type(subItem) == "table" then
+						return CNDT:GetTableSubstitution(subItem, "Spells_" .. suffix)
+					else
+						return strWrap(subItem)
+					end
+				end
+			end
+			return spells .. suffix
+		end
 	end,
 },
-
 {
-	src = "c.GCDReplacedNameFirst2",
+	src = "c%.OwnSpells(2?)(%.?[A-Za-z]*)",
 	rep = function(conditionData, conditionSettings, name, name2)
-
-		local name = TMW:GetSpells(name2).First
-		if name == "gcd" then
-			name = TMW.GCDSpell
+		return function(two, suffix)
+			local input = two == "2" and name2 or name
+			local spells = TMW:GetSpells(input, true)
+			if suffix:sub(1,1) == "." then
+				suffix = suffix:trim(".")
+				local subItem = spells[suffix]
+				if subItem then
+					if type(subItem) == "table" then
+						return CNDT:GetTableSubstitution(subItem, "Spells_" .. suffix)
+					else
+						return strWrap(subItem)
+					end
+				end
+			end
+			return spells .. suffix
 		end
-		return strWrap(name)
-	end,
-},{
-	src = "c.GCDReplacedNameFirst",
-	rep = function(conditionData, conditionSettings, name, name2)
-		local name = TMW:GetSpells(name).First
-		if name == "gcd" then
-			name = TMW.GCDSpell
-		end
-		return strWrap(name)
 	end,
 },
 
@@ -894,7 +874,8 @@ CNDT.Substitutions = {
 {
 	src = "LOWER(%b())",
 	rep = function() return strlower end,
-},}
+},
+}
 
 local conditionNameSettingProcessedCache = setmetatable(
 {}, {
@@ -931,7 +912,7 @@ function CNDT:GetConditionUnitSubstitution(unit)
 		-- so that in case the special unit doesn't map to anything,
 		-- we won't be passing nil to functions that don't accept nil.
 		substitution = 
-			CNDT:GetTableSubstitution(translatedUnits)
+			CNDT:GetTableSubstitution(translatedUnits, "Units")
 			.. "[1] or "
 			.. strWrap(firstOriginal)
 	else